level3_agentic_assistant/requirements.txt
requests        # to talk to OpenRouter API
python-dotenv   # to load API key safely from .env
